:name_badge: Rename class name
* 実装クラスとテストクラスのクラス名変更
* ディレクトリにネームスペースを併せる
